+++ ricaun.NUnit.Tests/NUnitTests.cs
@@ -325,7 +325,7 @@ namespace ricaun.NUnit.Tests
         [TestCase("*Abstract*", 4)]
         [TestCase("*.TestsCaseSource.*", 17)]
         [TestCase("*.InternalTests.*", 1)]
-        [TestCase("*", 78)]
+        [TestCase("*", 85)]
         public void TestAssembly_Filter(string testName, int numberOfTests)
         {
             TestEngineFilter.Add(testName);