+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `TestEngine.GetAssemblyMetadataAttributes` to get `AssemblyMetadataAttribute` from `assemblyFile`.
 - Add `AssemblyMetadataSerializable` to serialize `AssemblyMetadataAttribute` to `AssemblyMetadataSerializable`.
 - Add `AppDomainDisposable` internal class to create a new `AppDomain` and dispose it.
+- Add `TryChangeType` to try to convert `object` to `ParameterType`.
+### Tests
+- Add `SampleMetadata` and `SampleMetadataAttribute` test project.
+- Add `TestParameterCaseDouble` to test parameter with `double` type and parameter `int`. (Fix: #29)
+- Add `TestParameterByte`, `TestParameterShort` and `TestParameterDouble` tests for parameter converter.
 
 ## [1.4.0] / 2024-09-11
 ### Features